Added a GUI
This is a major change: I've went from a simple consol controller to a
GUI. I used QT as my GUI tools. It is set up so that new controllers can
be added quickly. This is V1 so there will be changes. I still need to
pass the output of the commands to the wireless transmission program.